Fix go get for packages in GOROOT
Previously go get would fail to run correctly when attempting to update a
package in GOROOT. This patch updates the build command to what the
upstream documentation recommends to use when building go from source.

This also adds a patch for a failing test due to strace output format
needing to have an updated regex.